Remove data generator
