+++ feed/src/main/java/com/apssouza/mytrade/feed/price/PriceDao.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public interface PriceDao {
-    void loadData(LocalDateTime start, LocalDateTime end);
 
     List<PriceDto> getPriceInterval(LocalDateTime start, LocalDateTime end);
